+++ packages/react-components/src/useConnect.ts
@@ -1,4 +1,4 @@
-import { useCallback, useState } from 'react';
+import { useMemo, useState } from 'react';
 import { WalletConnection, WalletConnector } from '@concordium/wallet-connectors';
 import { errorString } from './error';
 
@@ -7,9 +7,10 @@ import { errorString } from './error';
  */
 export interface Connect {
     /**
-     * Function for initiating a new connection. Any existing connection will not be automatically disconnected.
+     * Function for initiating a new connection using the provided connector or undefined if none was provided.
+     * Any existing connection will not be automatically disconnected by calling the function.
      */
-    connect: () => void;
+    connect: (() => void) | undefined;
 
     /**
      * Indicator on whether we're waiting for a connection to be established and approved.
@@ -25,7 +26,7 @@ export interface Connect {
 /**
  * Hook that exposes a function for initiating a connection on the provided {@link connector} and,
  * if successful, store the resulting connection in {@link setConnection}.
- * The hook also exposes the status of the connection progress and error if initiation failed.
+ * The hook also exposes the status of the connection progress and an error if initiation failed.
  * @param connector The connector from which new connections are to be initiated.
  * @param setConnection The setter function to which new connections are passed.
  */
@@ -35,21 +36,26 @@ export function useConnect(
 ): Connect {
     const [isConnecting, setIsConnecting] = useState(false);
     const [connectError, setConnectError] = useState('');
-    const connect = useCallback(() => {
+    const connect = useMemo(() => {
         if (!connector) {
-            throw new Error('no connector to connect from');
+            // There's no connector to connect to.
+            // Previously this would result in an exception being thrown by the returned function.
+            // Now we return 'undefined' instead to indicate that it isn't ready to be called.
+            return undefined;
         }
-        setIsConnecting(true);
-        connector
-            .connect()
-            .then((c) => {
-                if (c) {
-                    setConnection(c);
-                    setConnectError('');
-                }
-            })
-            .catch((e) => setConnectError(errorString(e)))
-            .finally(() => setIsConnecting(false));
+        return () => {
+            setIsConnecting(true);
+            connector
+                .connect()
+                .then((c) => {
+                    if (c) {
+                        setConnection(c);
+                        setConnectError('');
+                    }
+                })
+                .catch((e) => setConnectError(errorString(e)))
+                .finally(() => setIsConnecting(false));
+        };
     }, [connector, setConnection]);
     return { connect, isConnecting, connectError };
 }