+++ packages/react-components/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+-   Build as ES module to facilitate tree-shaking.
+
 ## [0.3.0] - 2023-06-04
 
 ### Added