+++ packages/react-components/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+-   Dependency on `@concordium/wallet-connectors` bumped to v0.5.0+.
+
 ## [0.4.0] - 2023-11-13
 
 ### Changed